feat(preference): new preference window

+++ src/renderer/components/LandingView.vue
@@ -60,6 +60,7 @@
 <script>
 import fs from 'fs';
 import { mapState, mapGetters } from 'vuex';
+import asyncStorage from '@/helpers/asyncStorage';
 import Titlebar from './Titlebar.vue';
 import Playlist from './LandingView/Playlist.vue';
 import NotificationBubble from './NotificationBubble.vue';
@@ -98,32 +99,38 @@ export default {
   },
   created() {
     // Get all data and show
-    this.infoDB.sortedResult('recent-played', 'lastOpened', 'prev')
-      .then((data) => {
-        const waitArray = [];
-        for (let i = 0; i < data.length; i += 1) {
-          const accessPromise = new Promise((resolve) => {
-            fs.access(data[i].path, fs.constants.F_OK, (err) => {
-              if (err) {
-                this.infoDB.delete('recent-played', data[i].quickHash);
-                resolve();
-              } else {
-                resolve(data[i]);
+    asyncStorage.get('preferences').then((data) => {
+      if (!data.deleteVideoHistoryOnExit) {
+        this.infoDB.sortedResult('recent-played', 'lastOpened', 'prev')
+          .then((data) => {
+            const waitArray = [];
+            for (let i = 0; i < data.length; i += 1) {
+              const accessPromise = new Promise((resolve) => {
+                fs.access(data[i].path, fs.constants.F_OK, (err) => {
+                  if (err) {
+                    this.infoDB.delete('recent-played', data[i].quickHash);
+                    resolve();
+                  } else {
+                    resolve(data[i]);
+                  }
+                });
+              });
+              waitArray.push(accessPromise);
+            }
+            return Promise.all(waitArray);
+          })
+          .then((data) => {
+            for (let i = 0; i < data.length; i += 1) {
+              if (data[i] === undefined) {
+                data.splice(i, 1);
               }
-            });
+            }
+            this.lastPlayedFile = data.slice(0, 9);
           });
-          waitArray.push(accessPromise);
-        }
-        return Promise.all(waitArray);
-      })
-      .then((data) => {
-        for (let i = 0; i < data.length; i += 1) {
-          if (data[i] === undefined) {
-            data.splice(i, 1);
-          }
-        }
-        this.lastPlayedFile = data.slice(0, 9);
-      });
+      } else {
+        this.infoDB.cleanData();
+      }
+    });
     this.$bus.$on('clean-lastPlayedFile', () => {
       // just for delete thumbnail display
       this.lastPlayedFile = [];
